+++ src/Resources/app/administration/src/main.js
@@ -10,6 +10,7 @@ import './module/extension/sw-sales-channel-modal-grid';
 import './module/extension/sw-settings-payment/components/swag-paypal-overview-card';
 import './module/extension/sw-settings-payment/sw-settings-payment-list';
 import './module/extension/sw-settings-payment/sw-settings-payment-detail';
+import './module/extension/sw-settings-shipping/sw-settings-shipping-detail';
 
 import './module/swag-paypal';
 import './module/swag-paypal-payment';